tests/wpt/web-platform-tests/html/semantics/embedded-content/the-canvas-element/2d.canvas.reference.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: 2d.canvas.reference</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>
<link rel="stylesheet" href="/common/canvas-tests.css">
<body class="show_output">

<h1>2d.canvas.reference</h1>
<p class="desc">CanvasRenderingContext2D.canvas refers back to its canvas</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("CanvasRenderingContext2D.canvas refers back to its canvas");
_addTest(function(canvas, ctx) {

_assertSame(ctx.canvas, canvas, "ctx.canvas", "canvas");


});
</script>
